% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{\enumerate{
\item 
}

downloadRedcapReport = function(redcapTokenName,redcapUrl,redcapReportId) :
Using the block of RedCap template code below, write a function that:
uses Sys.getenv() to read an API token called redcapTokenName from the
users’ .REnviron file.
queries redcapUrl to return the Redcap Report redcapReportId .
(Notice these are the data from
our simulated stroke study, now nicely and securely hosted on RedCap.)
returns the contents as a tibble}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the environment variable that stores the REDCap API token.}

\item{redcapUrl}{The URL of the REDCap.}

\item{redcapReportId}{The id of the report to be downloaded.}
}
\value{
A downloaded REDCap report tibble.
}
\description{
This function retrieve a report from REDCap using the specified API token, report ID, and REDCap API URL.
The api token should be stored in envrionment variable by using Sys.setenv().
The retrieved report returns in a tibble format.
}
\examples{
url = "https://redcap.emory.edu/api/"
Sys.setenv("REDCAP_API_TOKEN" = "6189879441F5C29A25245880677488BF")
print(downloadRedcapReport("REDCAP_API_TOKEN", url, '46524'))

}
