% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{\enumerate{
\item 
}

logLikBernoulli = function(data) Write a function that takes a vector like
data = c(1,0,0,0,1,1,1,) and calculates the parameter p that maximizes the log-likelihood
log(P(p|data)). Use a grid-based search with p in steps of 0.001.}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A numeric vector of 0s and 1s representing Bernoulli trials.}
}
\value{
The value of p that maximizes the log-likelihood for the given data.
}
\description{
This function estimates the maximum likelihood estimate for the parameter p of a Bernoulli distribution
using a grid search approach. It calculates the log-likelihood of observing the given data for
a sequence of p values ranging from 0 to 1 in steps of 0.001 and returns the p that maximizes the log-likelihood.
}
\examples{
data = c(1,0,0,0,1,1,1)
print(logLikBernoulli(data))

}
