% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{\enumerate{
\item 
}

minimumN = function(x1,x2) : Write a wrapper around pwr::pwr.t2n.test that takes either one ( x1 )
or two ( x2 ) samples of preliminary data and returns the minimum sample size needed for a t-test of the
null hypotheses that either μX1 == 0 or μX1 == μX2 with 80\% power at alpha=0.05.}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{Numeric vector of data for the first sample.}

\item{x2}{Numeric vector of data for the second sample.(Optional)}
}
\value{
The minimum sample size required to achieve 80\% power with a significance level of 0.05.
}
\description{
This function calculates the minimum sample size required to achieve a specified power
for a one-sample or two-sample t-test.
For a one-sample t-test, it tests if the mean of x1 is significantly different from zero.
For a two-sample t-test, it tests if the means of x1 and x2 are significantly different.
}
\examples{
set.seed(123)
x1 <- rnorm(20, mean = 5, sd = 1)
x2 <- rnorm(20, mean = 6, sd = 1)
minimumN(x1)
minimumN(x1, x2)

}
