% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{\enumerate{
\item 
}

pcApprox = function(x, npc) : Write a function that returns an approximation to the data x based on
npc PCs. (Note that the approximation should be rescaled and centered to match the original data).}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{A numeric matrix or data frame where rows are observations and columns are variables.}

\item{npc}{The number of principal components.}
}
\value{
A matrix containing the approximated data reconstructed from the first npc principal components.
}
\description{
This function performs PCA on the provided data matrix and uses the specified number of principal
components to approximate the original data. Then it reconstructs the data approximation from the
scores and loadings of the selected principal components.
Finally, it re-centers the approximated data using the original data's mean.
}
\examples{
data = mtcars
npc = 2
approx_data = pcApprox(data, npc)
print(data)
print(approx_data)

}
