% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{\enumerate{
\item 
}

standardizeNames = function(data) : Write a wrapper around dplyr::rename_with and
janitor::make_clean_names that converts the variables in a tibble data to "small_camel" case
(or another case if you like). The idea here is to have a reliable function that standardizes the variable
names in data you’re dealing with. This function should import elements of the janitor and snakecase
packages; Roxygen2 will handle these dependencies for you.}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A data frame whose column names you want to standardize.}
}
\value{
A data frame with standardized small camel case column names.
}
\description{
This function standardizes the variable names of a data frame to small camel case. It first cleans the names
to ensure they are syntactically valid, then converts them to small camel case using snakecase::to_any_case.
}
\examples{
test_data = tibble::tibble(`First name` = c("John"), `Last name` = c("Doe"))
standardized_data = standardizeNames(test_data)
print(standardized_data)

}
