% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{\enumerate{
\item 
}

unscale = function(x) : Write a function that takes a vector that has been put through scale and
reverses the centering/scaling, if any}
\usage{
unscale(x)
}
\arguments{
\item{x}{A numeric vector that may have been scaled and centered.}
}
\value{
A numeric vector with the scaling and centering reversed.
}
\description{
This function reverses the scaling and centering transformation applied to a numeric vector.
It checks if the vector x has attributes for 'scaled:center' and 'scaled:scale',
and if these attributes are present, it unscales the data by applying the inverse
of the original scaling and centering transformations.
}
\examples{
x = c(1, 2, 3, 4, 5)
unscale(scale(x))

}
